refactor: hoc code optimization
