+++ src/widgets/antd/list.jsx
@@ -1,3 +1,8 @@
+/**
+ * Updated by Tw93 on 2019-12-01.
+ * 数组组件
+ */
+
 import listHoc from '../../components/listHoc';
 import { Button } from 'antd';
