+++ src/widgets/antd/select.jsx
@@ -1,30 +1,9 @@
 import React from 'react';
 import { Select } from 'antd';
 
-const { Option } = Select;
+import selectHoc from '../../components/selectHoc';
 
 export default function select(p) {
-  const onChange = value => p.onChange(p.name, value);
-  return (
-    <Select
-      style={{ width: '100%' }}
-      {...p.options}
-      disabled={p.disabled}
-      value={p.value}
-      onChange={onChange}
-    >
-      {(p.schema.enum || []).map((val, index) => {
-        let option = p.schema.enumNames ? p.schema.enumNames[index] : val;
-        const isHtml = typeof option === 'string' && option[0] === '<';
-        if (isHtml) {
-          option = <span dangerouslySetInnerHTML={{ __html: option }} />;
-        }
-        return (
-          <Option value={val} key={index}>
-            {option}
-          </Option>
-        );
-      })}
-    </Select>
-  );
+  const FormSelect = selectHoc(p)(Select);
+  return <FormSelect />;
 }
